requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
Flask-Caching==2.1.0